+++ CHANGELOG.md
@@ -6,4 +6,5 @@
 - [Feature: delete implement ServiceInstance](https://github.com/Tencent/spring-cloud-tencent/pull/481)
 - [Upgrade owasp esapi's configuration](https://github.com/Tencent/spring-cloud-tencent/pull/492)
 - [Bugfix: update byte-buddy scope test to compile](https://github.com/Tencent/spring-cloud-tencent/pull/495)
+- [feature:add @ConditionalOnConfigReflectEnabled annotation](https://github.com/Tencent/spring-cloud-tencent/pull/496)
 - [Feature: zuul supports polaris router](https://github.com/Tencent/spring-cloud-tencent/pull/502)
