+++ spring-cloud-starter-tencent-polaris-config/src/test/java/com/tencent/cloud/polaris/config/adapter/PolarisPropertiesSourceAutoRefresherTest.java
@@ -26,7 +26,6 @@ import java.util.Map;
 
 import com.google.common.collect.Lists;
 import com.tencent.cloud.polaris.config.config.PolarisConfigProperties;
-import com.tencent.cloud.polaris.config.enums.RefreshType;
 import com.tencent.cloud.polaris.config.spring.property.PlaceholderHelper;
 import com.tencent.cloud.polaris.config.spring.property.SpringValue;
 import com.tencent.cloud.polaris.config.spring.property.SpringValueRegistry;
@@ -41,7 +40,6 @@ import org.mockito.junit.MockitoJUnitRunner;
 
 import org.springframework.beans.TypeConverter;
 import org.springframework.beans.factory.config.ConfigurableListableBeanFactory;
-import org.springframework.cloud.context.refresh.ContextRefresher;
 import org.springframework.context.ConfigurableApplicationContext;
 
 import static org.mockito.ArgumentMatchers.any;
@@ -49,7 +47,7 @@ import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.when;
 
 /**
- * test for {@link PolarisPropertySourceAutoRefresher}.
+ * test for {@link PolarisReflectConfigPropertyAutoRefresher}.
  *
  * @author lepdou 2022-06-11
  */
@@ -63,8 +61,6 @@ public class PolarisPropertiesSourceAutoRefresherTest {
 	private PolarisConfigProperties polarisConfigProperties;
 	@Mock
 	private PolarisPropertySourceManager polarisPropertySourceManager;
-	@Mock
-	private ContextRefresher contextRefresher;
 
 	@Mock
 	private SpringValueRegistry springValueRegistry;
@@ -74,10 +70,8 @@ public class PolarisPropertiesSourceAutoRefresherTest {
 
 	@Test
 	public void testConfigFileChanged() throws Exception {
-		PolarisPropertySourceAutoRefresher refresher = new PolarisPropertySourceAutoRefresher(polarisConfigProperties,
-				polarisPropertySourceManager, springValueRegistry, placeholderHelper, contextRefresher);
-
-		when(polarisConfigProperties.getRefreshType()).thenReturn(RefreshType.REFLECT);
+		PolarisReflectConfigPropertyAutoRefresher refresher = new PolarisReflectConfigPropertyAutoRefresher(polarisConfigProperties,
+				polarisPropertySourceManager, springValueRegistry, placeholderHelper);
 		ConfigurableApplicationContext applicationContext = mock(ConfigurableApplicationContext.class);
 		ConfigurableListableBeanFactory beanFactory = mock(ConfigurableListableBeanFactory.class);
 		TypeConverter typeConverter = mock(TypeConverter.class);
