+++ spring-cloud-tencent-commons/src/test/java/com/tencent/cloud/common/metadata/StaticMetadataManagerTest.java
@@ -38,7 +38,7 @@ import static org.mockito.Mockito.when;
 
 
 /**
- * test for {@link StaticMetadataManager}
+ * test for {@link StaticMetadataManager}.
  *@author lepdou 2022-06-27
  */
 @RunWith(MockitoJUnitRunner.class)
